refactor(CampaignCard): move CampaignStatus into separate file
